+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - New `typo3_mysql` credential formatter for easy formatting of the typo3 database connection
+- Allow to register your own credential formatter with the DI tag `bartacus.platformsh.credential_formatter`
 
 ### Breaking
 - Updated to use `platformsh/config-reader` @ `^2.1`